fix: ranker
